+++ src/main/java/com/example/springlogin/member/controller/MemberController.java
@@ -1,5 +1,6 @@
 package com.example.springlogin.member.controller;
 
+import com.example.springlogin.member.controller.request.LoginRequest;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Controller;
