+++ src/main/java/com/example/springlogin/member/controller/request/JoinRequest.java
@@ -1,4 +1,4 @@
-package com.example.springlogin.member.controller;
+package com.example.springlogin.member.controller.request;
 
 import lombok.Builder;
 import lombok.Data;
