+++ src/main/java/com/example/springlogin/member/service/MemberService.java
@@ -1,6 +1,8 @@
 package com.example.springlogin.member.service;
 
 import com.example.springlogin.member.domain.Member;
+import com.example.springlogin.member.service.param.JoinParam;
+import com.example.springlogin.member.service.param.LoginParam;
 
 import java.util.Optional;
