+++ src/main/java/com/example/springlogin/member/service/param/LoginParam.java
@@ -1,13 +1,11 @@
-package com.example.springlogin.member.service;
-
+package com.example.springlogin.member.service.param;
 
 import lombok.Builder;
 import lombok.Data;
 
-
 @Data
 @Builder
-public class JoinParam {
+public class LoginParam {
     private String email;
     private String password;
 }
